moved out events, inputs and controllers to editor
