+++ Tetragrama/Components/HierarchyViewUIComponent.h
@@ -1,7 +1,8 @@
 #pragma once
 #include <EditorCameraController.h>
 #include <Message.h>
-#include <ZEngine/ZEngine.h>
+#include <ZEngine/Components/UIComponent.h>
+#include <future>
 #include <mutex>
 #include <string>
 
@@ -20,14 +21,14 @@ namespace Tetragrama::Components
         void RenderGuizmo();
         void RenderSceneNodeTree(int node_identifier);
 
-        std::future<void> EditorCameraAvailableMessageHandlerAsync(Messengers::GenericMessage<ZEngine::Ref<EditorCameraController>>&);
+        std::future<void> EditorCameraAvailableMessageHandlerAsync(Messengers::GenericMessage<ZEngine::Ref<Controllers::EditorCameraController>>&);
 
     private:
-        ImGuiTreeNodeFlags                       m_node_flag;
-        bool                                     m_is_node_opened{false};
-        int                                      m_selected_node_identifier{-1};
-        int                                      m_gizmo_operation{-1};
-        std::mutex                               m_mutex;
-        ZEngine::WeakRef<EditorCameraController> m_active_editor_camera;
+        ImGuiTreeNodeFlags                                    m_node_flag;
+        bool                                                  m_is_node_opened{false};
+        int                                                   m_selected_node_identifier{-1};
+        int                                                   m_gizmo_operation{-1};
+        std::mutex                                            m_mutex;
+        ZEngine::WeakRef<Controllers::EditorCameraController> m_active_editor_camera;
     };
 } // namespace Tetragrama::Components
