+++ ZEngine/ZEngine/Event/KeyEvent.h
@@ -21,4 +21,53 @@ namespace ZEngine::Event
     protected:
         ZENGINE_KEYCODE m_keycode;
     };
+
+    class KeyPressedEvent : public KeyEvent
+    {
+    public:
+        KeyPressedEvent(ZENGINE_KEYCODE key, int repeat_count) : KeyEvent(key), m_repeat_count(repeat_count) {}
+
+        EVENT_TYPE(KeyPressed)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("KeyPressedEvent : {0}, repeated count : {1}", m_keycode, m_repeat_count);
+        }
+
+    protected:
+        int m_repeat_count{0};
+    };
+
+    class KeyReleasedEvent : public KeyEvent
+    {
+    public:
+        KeyReleasedEvent(ZENGINE_KEYCODE key) : KeyEvent(key) {}
+
+        EVENT_TYPE(KeyReleased)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("KeyReleasedEvent : {0}", m_keycode);
+        }
+    };
 } // namespace ZEngine::Event
