+++ ZEngine/ZEngine/Event/MouseEvent.h
@@ -23,4 +23,124 @@ namespace ZEngine::Event
     protected:
         ZENGINE_KEYCODE m_button{0};
     };
+
+    class MouseButtonPressedEvent : public MouseEvent
+    {
+    public:
+        MouseButtonPressedEvent(ZENGINE_KEYCODE button) : MouseEvent(button) {}
+
+        EVENT_TYPE(MouseButtonPressed)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("MouseButtonPressedEvent : {0}", m_button);
+        }
+    };
+
+    class MouseButtonReleasedEvent : public MouseEvent
+    {
+    public:
+        explicit MouseButtonReleasedEvent(ZENGINE_KEYCODE button) : MouseEvent(button) {}
+
+        EVENT_TYPE(MouseButtonReleased)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("MouseButtonReleasedEvent : {0}", m_button);
+        }
+    };
+
+    class MouseButtonMovedEvent : public MouseEvent
+    {
+    public:
+        MouseButtonMovedEvent(double xpos, double ypos) : m_xpos(xpos), m_ypos(ypos) {}
+
+        double GetPosX() const
+        {
+            return m_xpos;
+        }
+
+        double GetPosY() const
+        {
+            return m_ypos;
+        }
+
+        EVENT_TYPE(MouseMoved)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("MouseButtonMovedEvent");
+        }
+
+    private:
+        double m_xpos{0};
+        double m_ypos{0};
+    };
+
+    class MouseButtonWheelEvent : public MouseEvent
+    {
+    public:
+        MouseButtonWheelEvent(double offset_x, double offset_y) : m_offset_x(offset_x), m_offset_y(offset_y) {}
+
+        double GetOffetX() const
+        {
+            return m_offset_x;
+        }
+
+        double GetOffetY() const
+        {
+            return m_offset_y;
+        }
+
+        EVENT_TYPE(MouseWheel)
+
+        virtual EventType GetType() const override
+        {
+            return GetStaticType();
+        }
+
+        virtual int GetCategory() const override
+        {
+            return GetStaticCategory();
+        }
+
+        virtual std::string ToString() const override
+        {
+            return fmt::format("MouseButtonWheelEvent : {0}", m_button);
+        }
+
+    private:
+        double m_offset_x{0};
+        double m_offset_y{0};
+    };
 } // namespace ZEngine::Event
