--- ZEngine/ZEngine/Inputs/ITextInputEventCallback.h
@@ -1,14 +1,0 @@
-#pragma once
-#include <Event/TextInputEvent.h>
-
-namespace ZEngine::Inputs
-{
-
-    struct ITextInputEventCallback
-    {
-        ITextInputEventCallback()  = default;
-        ~ITextInputEventCallback() = default;
-
-        virtual bool OnTextInputRaised(ZEngine::Event::TextInputEvent&) = 0;
-    };
-} // namespace ZEngine::Inputs
