+++ ZEngine/ZEngine/Window/CoreWindow.h
@@ -8,15 +8,11 @@
 #include <Event/CoreEvent.h>
 #include <Event/EventDispatcher.h>
 #include <Event/TextInputEvent.h>
-#include <Event/WindowClosedEvent.h>
-#include <Event/WindowResizedEvent.h>
-#include <Inputs/IKeyboardEventCallback.h>
-#include <Inputs/IMouseEventCallback.h>
-#include <Inputs/ITextInputEventCallback.h>
+#include <Event/WindowEvent.h>
+#include <Inputs/IInputEventCallback.h>
 #include <Layers/Layer.h>
 #include <Layers/LayerStack.h>
 #include <Rendering/Swapchain.h>
-#include <Window/ICoreWindowEventCallback.h>
 #include <Window/WindowConfiguration.h>
 #include <Window/WindowProperty.h>
 
@@ -29,7 +25,16 @@ namespace ZEngine::Layers
 namespace ZEngine::Window
 {
 
-    class CoreWindow : public Helpers ::RefCounted,
+    struct ICoreWindowEventCallback
+    {
+        virtual bool OnWindowClosed(Event::WindowClosedEvent&)       = 0;
+        virtual bool OnWindowResized(Event::WindowResizedEvent&)     = 0;
+        virtual bool OnWindowMinimized(Event::WindowMinimizedEvent&) = 0;
+        virtual bool OnWindowMaximized(Event::WindowMaximizedEvent&) = 0;
+        virtual bool OnWindowRestored(Event::WindowRestoredEvent&)   = 0;
+    };
+
+    class CoreWindow : public Helpers::RefCounted,
                        public Inputs::IKeyboardEventCallback,
                        public Inputs::IMouseEventCallback,
                        public Inputs::ITextInputEventCallback,
