+++ ZEngine/ZEngine/ZEngine.h
@@ -8,8 +8,6 @@
 #include "Core/TimeStep.h"
 #include "Core/Utility.h"
 #include "Engine.h"
-#include "Event/EventCategory.h"
-#include "Event/EventType.h"
 #include "Inputs/IDevice.h"
 #include "Inputs/KeyCode.h"
 #include "Inputs/KeyCodeDefinition.h"
